Changing from Classe to Character
