--- Assets/Scripts/Classes/Classe.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 89eae6946d4b6594e878732d18365a55
-timeCreated: 1506023872
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
